Refactored name of file to exercise_workout_view to better describle its purpose
